refactor: rename plugin.js to index.js
