+++ src/index.js
@@ -9,7 +9,7 @@ module.exports = Filer = {
   path: require('./path.js'),
   Errors: require('./errors.js'),
   Shell: require('./shell/shell.js'),
-  FilerWebpackPlugin: require('./webpack-plugin/plugin'),
+  FilerWebpackPlugin: require('./webpack-plugin'),
 };
 
 // Add a getter for the `fs` instance, which returns
