+++ src/webpack-plugin/index.js
